clean up background animation
