+++ webinterface/src/app/app.component.html
@@ -1,27 +1,20 @@
 <app-display-map>
 
 
-
   <div class="buttons">
 
     <h2>J+S-Marschzeittabellen automatisiert generieren</h2>
 
-    <ng-container *ngIf="!pending else loading">
-      <input type="file" id="uploader" (input)="gpx_uploaded()"
-             id="uploader" accept=".gpx">
+    <ng-container *ngIf="!pending else pendingText">
+      <input type="file" id="uploader" id="uploader" accept=".gpx">
 
       <button id="export-button" (click)="download_map()">Generieren lassen</button>
     </ng-container>
 
-    <ng-template #loading>
-      Marschzeittabelle wird erstellt, wir bitten um Geduld.
+    <ng-template #pendingText>
+      <p> Marschzeittabelle wird erstellt, wir bitten um Geduld.</p>
+      <p> <b>Status:</b> {{status}} <br> <b>Message:</b> {{status_message}} </p>
     </ng-template>
-
-
-    <ng-container *ngIf="showDownloadLink">
-      <br/>
-      <a href="http://localhost:5000/download/{{uuid}}">Download Data</a>
-    </ng-container>
 
   </div>
 
@@ -30,5 +23,3 @@
 
 
 <app-footer></app-footer>
-
-
