+++ webinterface/src/app/app.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {environment} from "../environments/environment";
-import * as gpxParser from 'gpxparser';
+import {MapAnimatorService} from "./services/map-animator.service";
 
 
 @Component({
@@ -10,24 +10,24 @@ import * as gpxParser from 'gpxparser';
 })
 export class AppComponent implements OnInit {
 
+  static baseURL = environment.API_URL;
+
+  title = 'automatic-walk-time-tables';
+  pending = false;
+  uuid: string = '';
+  status_message: string = '';
+  status: string = '';
+
+  constructor(private mapAnimator: MapAnimatorService) {
+  }
 
   ngOnInit(): void {
 
   }
 
 
-  static baseURL = environment.API_URL;
-
-
-  title = 'automatic-walk-time-tables';
-  pending = false;
-  showDownloadLink = false;
-  uuid: string = '';
-
-
   download_map() {
 
-    this.showDownloadLink = false;
     this.pending = true;
     const uploaderElement = (document.getElementById('uploader') as HTMLInputElement);
 
@@ -36,6 +36,10 @@ export class AppComponent implements OnInit {
 
     // @ts-ignore
     let gpx_file: File = uploaderElement.files[0];
+
+    // Start Animation
+    this.mapAnimator.add_gpx_file(gpx_file);
+
     let formData = new FormData();
 
     formData.append("file", gpx_file);
@@ -55,30 +59,32 @@ export class AppComponent implements OnInit {
         this.uuid = response['uuid'];
         console.log(this.uuid)
 
-        this.download_data()
-
-        AppComponent.get_status_updates(this.uuid)
+        this.get_status_updates()
 
       })
       .catch(console.log)
 
   }
 
-  static get_status_updates(uuid: string) {
+  get_status_updates() {
 
     console.log('Request a status update.')
 
     const baseURL = AppComponent.baseURL + "status/"
 
-    fetch(baseURL + uuid)
+    fetch(baseURL + this.uuid)
       .then(response => response.json())
       .then(res => {
-        console.log(res);
 
-        if (['error', 'finished'].includes(res.status))
+        this.status = res.status;
+        this.status_message = res.message;
+
+        if (['error', 'finished'].includes(res.status)) {
+          this.download_data()
           return;
+        }
 
-        setTimeout(() => this.get_status_updates(uuid), 500)
+        setTimeout(() => this.get_status_updates(), 500)
       });
 
   }
@@ -86,11 +92,9 @@ export class AppComponent implements OnInit {
   download_data() {
 
     this.pending = false;
-    this.showDownloadLink = true;
+    window.location.href = "http://localhost:5000/download/" + this.uuid;
 
   }
 
-  gpx_uploaded() {
 
-  }
 }
