+++ webinterface/src/app/display-map/display-map.component.scss
@@ -1,11 +1,10 @@
-#map-canvas{
+#map-canvas {
   width: 100%;
   height: 100%;
   position: absolute;
   top: 0;
   left: 0;
   overflow: hidden;
-  // opacity: 0.2;
   z-index: -10;
 }
 
@@ -20,6 +19,10 @@
 
 .content {
 
-  margin: 180px 25px 25px;
+  margin: 180px auto auto;
+
+  background-color: #ffffffa0;
+  padding: 150px 50px;
+  max-width: 700px;
 
 }
